reverted file that should not have been changed in [25852]; refs #13484; !strict
git-svn-id: http://svn.dojotoolkit.org/src/util/trunk@25855 560b804f-0ae3-0310-86f3-f6aa0a117693

+++ buildscripts/build.sh
@@ -1,8 +1,8 @@
 #!/bin/sh
 
-#/if which -s node
-#then
-#   node ../../dojo/dojo.js load=build "$@"
-#else
+if which -s node
+then
+   node ../../dojo/dojo.js load=build "$@"
+else
    java -Xms256m -Xmx256m  -cp ../shrinksafe/js.jar:../closureCompiler/compiler.jar:../shrinksafe/shrinksafe.jar org.mozilla.javascript.tools.shell.Main  ../../dojo/dojo.js baseUrl=../../dojo load=build "$@"
-#fi
+fi
